pyapp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.7
langchain-openai==0.0.5
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
redis==5.0.1
elasticsearch==8.11.0
nebula3-python==3.8.0
pymilvus==2.3.4
mcp==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
pyyaml==6.0.1
loguru==0.7.2
requests==2.31.0
httpx==0.25.2
asyncio==3.4.3
aiofiles==23.2.1
python-dotenv==1.0.0